Web_Development/requirements.txt
Django==2.2.3
django-crispy-forms==1.7.2
django-mdeditor==0.1.14
Markdown==3.1.1
Pillow==6.1.0
Pygments==2.4.2
pytz==2019.1
sqlparse==0.3.0
